pyproject.toml
[project]
name = "legal-vl-2-0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.65.0",
    "pymupdf>=1.25.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "tqdm>=4.67.3",
]
